+++ computer_vision/to_json.py
@@ -10,6 +10,3 @@ def to_json(dirpath: str):
 
             with open(os.path.join(dirpath, filename.split(".")[0] + ".json"), "w") as file:
                 json.dump(dictionary, file)
-
-if __name__ == "__main__":
-    to_json("/home/otokonokage/Documents/github/deep_learning_python/computer_vision/yolo/yolov6/model_saves/modified_coco2017/asu-4862/")