+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this repository will be documented in this file.
 
+## 2026-06-02 - Correction Data Augmentation
+- Wrong box format for data augmentation.
+
 ## 2026-06-02 Editing README.md and other stuff
 
 ## 2026-05-06 Kaggle corrections again